GSF.Core\FuzzyStrings: Fix issue where a double was being assigned the quotient of 2 integers by casting the denominators as doubles

+++ Source/Libraries/GSF.Core/FuzzyStrings/ApproximatelyEquals.cs
@@ -47,7 +47,7 @@ namespace GSF.FuzzyStrings
             {
                 if (source.Length == target.Length)
                 {
-                    comparisonResults.Add(source.HammingDistance(target) / target.Length);
+                    comparisonResults.Add(source.HammingDistance(target) / (double)target.Length);
                 }
             }
